Move TimeControl on game/
