+++ src/game/cTimeCounter.h
