+++ src/game/cTimeManager.h
